+++ spec/password_blocklist/checker_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe PasswordBlacklist::Checker do
+describe PasswordBlocklist::Checker do
   subject(:checker) { described_class.new }
 
   let(:preloaded) { described_class.new(:xs) }
@@ -25,72 +25,72 @@ describe PasswordBlacklist::Checker do
     end
   end
 
-  describe '#blacklisted?' do
+  describe '#blocklisted?' do
     it 'returns true if the provided password is in the blocklist' do
-      expect(checker.blacklisted?('pokemon')).to be(true)
+      expect(checker.blocklisted?('pokemon')).to be(true)
     end
 
     it 'returns false if the provided password is not in the blocklist' do
-      expect(checker.blacklisted?('CorrectHorseBatteryStaple')).to be(false)
+      expect(checker.blocklisted?('CorrectHorseBatteryStaple')).to be(false)
     end
 
     describe 'using differente sized blocklists' do
       it 'supports an extra small list' do
-        expect(checker.blacklisted?('pokemon', :xs)).to be(true)
-        expect(checker.blacklisted?('assassins', :xs)).to be(false)
+        expect(checker.blocklisted?('pokemon', :xs)).to be(true)
+        expect(checker.blocklisted?('assassins', :xs)).to be(false)
       end
 
       it 'supports a small list' do
-        expect(checker.blacklisted?('assassins', :sm)).to be(true)
-        expect(checker.blacklisted?('blackeyedpeas', :sm)).to be(false)
+        expect(checker.blocklisted?('assassins', :sm)).to be(true)
+        expect(checker.blocklisted?('blackeyedpeas', :sm)).to be(false)
       end
 
       it 'supports a medium list' do
-        expect(checker.blacklisted?('blackeyedpeas', :md)).to be(true)
-        expect(checker.blacklisted?('deduplication', :md)).to be(false)
+        expect(checker.blocklisted?('blackeyedpeas', :md)).to be(true)
+        expect(checker.blocklisted?('deduplication', :md)).to be(false)
       end
 
       it 'supports a large list' do
-        expect(checker.blacklisted?('deduplication', :lg)).to be(true)
-        expect(checker.blacklisted?('amakingsortofbutton', :lg)).to be(false)
+        expect(checker.blocklisted?('deduplication', :lg)).to be(true)
+        expect(checker.blocklisted?('amakingsortofbutton', :lg)).to be(false)
       end
 
       it 'supports an extra large list' do
-        expect(checker.blacklisted?('amakingsortofbutton', :xl)).to be(true)
-        expect(checker.blacklisted?('CorrectHorseBattleStaple', :xl))
+        expect(checker.blocklisted?('amakingsortofbutton', :xl)).to be(true)
+        expect(checker.blocklisted?('CorrectHorseBattleStaple', :xl))
           .to be(false)
       end
 
       it 'raise an error for unknown list size' do
-        expect { checker.blacklisted?('pokemon', :abc) }
-          .to raise_error(PasswordBlacklist::Checker::UnknownListSizeError)
+        expect { checker.blocklisted?('pokemon', :abc) }
+          .to raise_error(PasswordBlocklist::Checker::UnknownListSizeError)
       end
     end
 
     it 'ignores password case sensitivity' do
-      expect(checker.blacklisted?('PokeMon', :sm)).to be(true)
+      expect(checker.blocklisted?('PokeMon', :sm)).to be(true)
     end
 
     it 'escapes regular expression characters' do
-      expect(checker.blacklisted?(')(', :sm)).to be(false)
-      expect(checker.blacklisted?('.*', :sm)).to be(false)
+      expect(checker.blocklisted?(')(', :sm)).to be(false)
+      expect(checker.blocklisted?('.*', :sm)).to be(false)
     end
 
     it 'supports strings for list size' do
-      expect(checker.blacklisted?('password', 'sm')).to be(true)
+      expect(checker.blocklisted?('password', 'sm')).to be(true)
     end
 
     describe 'in-memory list' do
       it 'loaded when empty' do
-        checker.blacklisted?('pokemon')
+        checker.blocklisted?('pokemon')
 
         expect(File).to have_received(:read)
       end
 
       it 'loaded and retained' do
-        checker.blacklisted?('pokemon')
+        checker.blocklisted?('pokemon')
 
-        checker.blacklisted?('password')
+        checker.blocklisted?('password')
 
         expect(File).to have_received(:read).once
       end
@@ -98,7 +98,7 @@ describe PasswordBlacklist::Checker do
       it 'is used if preloaded' do
         preloaded
 
-        preloaded.blacklisted?('pokemon')
+        preloaded.blocklisted?('pokemon')
 
         expect(File).to have_received(:read).once
       end
@@ -106,7 +106,7 @@ describe PasswordBlacklist::Checker do
       it 'is used if the list is the same' do
         preloaded
 
-        preloaded.blacklisted?('pokemon', :xs)
+        preloaded.blocklisted?('pokemon', :xs)
 
         expect(File).to have_received(:read).once
       end
@@ -114,7 +114,7 @@ describe PasswordBlacklist::Checker do
       it 'is updated when a different sized list is specificed' do
         preloaded
 
-        preloaded.blacklisted?('pokemon', :xl)
+        preloaded.blocklisted?('pokemon', :xl)
 
         expect(File).to have_received(:read).twice
       end
@@ -122,9 +122,9 @@ describe PasswordBlacklist::Checker do
       it 'is updated and retained when the list has changed' do
         preloaded
 
-        preloaded.blacklisted?('pokemon', :xl)
-        preloaded.blacklisted?('password', :xl)
-        preloaded.blacklisted?('querty')
+        preloaded.blocklisted?('pokemon', :xl)
+        preloaded.blocklisted?('password', :xl)
+        preloaded.blocklisted?('querty')
 
         expect(File).to have_received(:read).twice
       end